Add custom CSS (fix accent colours, fonts)
Also removed a couple of leftovers of the previous theme.
This commit is a manual fix for https://github.com/curtiscde/hugo-theme-massively/issues/116
I will try and implement a more general solution to allow colour customisation in a PR soon (or at least an adaptation of this custom.css+htmlhead.custom.html as a better example).